requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0; sys_platform != "win32"
uvicorn==0.24.0; sys_platform == "win32"
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
numpy>=1.26.4
requests>=2.31.0
httpx>=0.24.3,<0.28
cachetools>=5.3.0
zstandard>=0.22.0